`http_serde_priv` module
