+++ src/interface/config.rs
@@ -14,6 +14,8 @@ use crate::{
     interface::template::Template,
 };
 
+use super::helper::http_serde_priv;
+
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
 #[serde(deny_unknown_fields, rename_all = "kebab-case")]
 #[serde(bound = "Q: Configuration, P: Configuration")]
@@ -431,127 +433,6 @@ pub mod destinations {
             D: Deserializer<'de>,
         {
             T::deserialize(deserializer).map(Transpose::transpose)
-        }
-    }
-}
-
-// `http` do not support serde https://github.com/hyperium/http/pull/631
-pub(crate) mod http_serde_priv {
-    use std::ops::{Deref, DerefMut};
-
-    use super::*;
-
-    #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-    pub struct Method(#[serde(with = "http_serde::method")] pub http::Method);
-    impl From<http::Method> for Method {
-        fn from(m: http::Method) -> Self {
-            Self(m)
-        }
-    }
-    impl Deref for Method {
-        type Target = http::Method;
-        fn deref(&self) -> &Self::Target {
-            &self.0
-        }
-    }
-    impl DerefMut for Method {
-        fn deref_mut(&mut self) -> &mut Self::Target {
-            &mut self.0
-        }
-    }
-
-    #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-    pub struct StatusCode(#[serde(with = "http_serde::status_code")] pub http::StatusCode);
-    impl From<http::StatusCode> for StatusCode {
-        fn from(s: http::StatusCode) -> Self {
-            Self(s)
-        }
-    }
-    impl Deref for StatusCode {
-        type Target = http::StatusCode;
-        fn deref(&self) -> &Self::Target {
-            &self.0
-        }
-    }
-    impl DerefMut for StatusCode {
-        fn deref_mut(&mut self) -> &mut Self::Target {
-            &mut self.0
-        }
-    }
-
-    #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-    pub struct Uri(#[serde(with = "http_serde::uri")] pub http::Uri);
-    impl From<http::Uri> for Uri {
-        fn from(u: http::Uri) -> Self {
-            Self(u)
-        }
-    }
-    impl Deref for Uri {
-        type Target = http::Uri;
-        fn deref(&self) -> &Self::Target {
-            &self.0
-        }
-    }
-    impl DerefMut for Uri {
-        fn deref_mut(&mut self) -> &mut Self::Target {
-            &mut self.0
-        }
-    }
-
-    #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-    pub struct Version(#[serde(with = "http_serde::version")] pub http::Version);
-    impl From<http::Version> for Version {
-        fn from(v: http::Version) -> Self {
-            Self(v)
-        }
-    }
-    impl Deref for Version {
-        type Target = http::Version;
-        fn deref(&self) -> &Self::Target {
-            &self.0
-        }
-    }
-    impl DerefMut for Version {
-        fn deref_mut(&mut self) -> &mut Self::Target {
-            &mut self.0
-        }
-    }
-
-    #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-    pub struct HeaderMap(#[serde(with = "http_serde::header_map")] pub http::HeaderMap);
-    impl From<http::HeaderMap> for HeaderMap {
-        fn from(m: http::HeaderMap) -> Self {
-            Self(m)
-        }
-    }
-    impl Deref for HeaderMap {
-        type Target = http::HeaderMap;
-        fn deref(&self) -> &Self::Target {
-            &self.0
-        }
-    }
-    impl DerefMut for HeaderMap {
-        fn deref_mut(&mut self) -> &mut Self::Target {
-            &mut self.0
-        }
-    }
-
-    #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-    pub struct Authority(#[serde(with = "http_serde::authority")] pub http::uri::Authority);
-    impl From<http::uri::Authority> for Authority {
-        fn from(a: http::uri::Authority) -> Self {
-            Self(a)
-        }
-    }
-    impl Deref for Authority {
-        type Target = http::uri::Authority;
-        fn deref(&self) -> &Self::Target {
-            &self.0
-        }
-    }
-    impl DerefMut for Authority {
-        fn deref_mut(&mut self) -> &mut Self::Target {
-            &mut self.0
         }
     }
 }
