+++ src/interface/mod.rs
@@ -1,4 +1,5 @@
 pub mod command;
 pub mod config;
+pub mod helper;
 pub mod report;
 pub mod template;
